requirements.txt
numpy>=1.26
opencv-python>=4.9
Pillow>=10.2
pillow-heif>=0.15
PyMuPDF>=1.23
